pyproject.toml
[project]
name = "notifier-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aio-pika>=9.5.7",
    "fastapi>=0.120.2",
    "pydantic-settings>=2.11.0",
    "python-telegram-bot>=22.5",
    "uvicorn>=0.38.0",
]
